move assignment table to its own component
